+++ project/BuildPlugin.scala
@@ -1,5 +1,5 @@
-import de.heikoseeberger.sbtheader.HeaderPlugin.autoImport.{HeaderLicense, headerLicense}
-import de.heikoseeberger.sbtheader.{HeaderPlugin, License}
+import de.heikoseeberger.sbtheader.HeaderPlugin
+import de.heikoseeberger.sbtheader.HeaderPlugin.autoImport.headerLicense
 import sbt.Keys._
 import sbt._
 import sbt.plugins.JvmPlugin
@@ -34,7 +34,7 @@ object BuildPlugin extends AutoPlugin {
     run / fork               := true,
     Test / fork              := true,
     Test / parallelExecution := false,
-    headerLicense            := Some(headerIOLicense),
+    headerLicense            := Some(Header.headerIOLicense),
     Compile / console / scalacOptions ~= (_.filterNot(
       Set("-Xfatal-warnings", "-Ywarn-unused:imports")
     )),
@@ -47,34 +47,4 @@ object BuildPlugin extends AutoPlugin {
     Test / testOptions += Tests.Argument(TestFrameworks.JUnit, "-v", "-a")
   ) ++ SonatypePublish.projectSettings
 
-  /**
-   * SBT Header Plugin
-   */
-
-  lazy val headerText: String =
-    """|MIT License
-       |
-       |Copyright (c) 2024 Rafael Fernandez
-       |
-       |Permission is hereby granted, free of charge, to any person obtaining a copy
-       |of this software and associated documentation files (the "Software"), to deal
-       |in the Software without restriction, including without limitation the rights
-       |to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-       |copies of the Software, and to permit persons to whom the Software is
-       |furnished to do so, subject to the following conditions:
-       |
-       |The above copyright notice and this permission notice shall be included in all
-       |copies or substantial portions of the Software.
-       |
-       |THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-       |IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-       |FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-       |AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-       |LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-       |OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-       |SOFTWARE.
-       |""".stripMargin
-
-  lazy val headerIOLicense: License.Custom =
-    HeaderLicense.Custom(headerText)
 }
