+++ project/plugins.sbt
@@ -1,4 +1,6 @@
 /** Compiler */
+//addSbtPlugin("org.typelevel"      % "sbt-tpolecat"       % "0.5.2")
+addSbtPlugin("org.scoverage"      % "sbt-scoverage"      % "2.3.0")
 addCompilerPlugin("com.olegpy"   %% "better-monadic-for" % "0.3.1")
 addCompilerPlugin("org.typelevel" % "kind-projector"     % "0.13.2" cross CrossVersion.full)
